Set the icon_size property when creating the science icons
While dynamically creating all the sciences, the original icon is cleared and a new one is defined with the icons array. Problem is: the icon_size property is not set. This can lead to a startup error if some other mod adds another icon to this array, say an overlay, which has the icon_size set. Seen with GDIW which tried to create a mirrored fluid recipe, adding its rotated marker to the array.

I don't know if this is also needed for the technology, but for consistency.